napravito del

+++ src/actions/DeleteAction.java
@@ -5,7 +5,9 @@ import gui.MainFrame;
 import gui.NorthTablePanel;
 import resource.data.Row;
 
+import javax.swing.*;
 import java.awt.event.ActionEvent;
+import java.util.HashMap;
 import java.util.Map;
 
 public class DeleteAction extends MyAbstractAction {
@@ -17,14 +19,22 @@ public class DeleteAction extends MyAbstractAction {
 
     @Override
     public void actionPerformed(ActionEvent e) {
-        System.out.println(" DEL ");
+        //System.out.println(" DEL ");
         //TODO greska ako nije selektovano
 
 
         NorthTablePanel ntp = (NorthTablePanel)MainFrame.getInstance().getNorthTab().getSelectedComponent();
+        if(ntp.getjTable().getSelectedRow() == -1){
+            JOptionPane.showMessageDialog(null, "Row not Selected", "Error", JOptionPane.ERROR_MESSAGE);
+            return;
+        }
+        int x = JOptionPane.showConfirmDialog(MainFrame.getInstance(), "Confirm the action delete");
+        if(x>0)
+            return;
         Row currentRow = ntp.getTableModel().getRows().get(ntp.getjTable().getSelectedRow());
 
 
+        /*
         StringBuilder stringBuilder = new StringBuilder();
         stringBuilder.append("DELETE FROM ");
         stringBuilder.append(currentRow.getName());
@@ -38,12 +48,17 @@ public class DeleteAction extends MyAbstractAction {
                 stringBuilder.append(" AND ");
             }
             stringBuilder.append(entry);
+            stringBuilder.append(entry.getKey()+"=");
+            stringBuilder.append("'"+entry.getValue().toString()+"'");
 
         }
         stringBuilder.append(";");
-        System.out.println(stringBuilder);
+        System.out.println(stringBuilder);*/
         String konst = "DELETE FROM REGIONS WHERE region_id='6.0' AND region_name='Westeros';";
-        MainFrame.getInstance().getAppCore().executeQuery(konst,null,((NorthTablePanel) MainFrame.getInstance().getNorthTab().getSelectedComponent()).getTableModel(),ntp.getEntity().getName());
+        //MainFrame.getInstance().getAppCore().executeQuery(konst,null,((NorthTablePanel) MainFrame.getInstance().getNorthTab().getSelectedComponent()).getTableModel(),ntp.getEntity().getName());
+        MainFrame.getInstance().getAppCore().deleteToTable((HashMap<String, Object>) currentRow.getFields(),
+                ((NorthTablePanel) MainFrame.getInstance().getNorthTab().getSelectedComponent()).getTableModel(),
+                ntp.getEntity().getName());
 
 
     }
